+++ demo/demo2.cpp
@@ -43,7 +43,7 @@ int main() {
         if (SUCCEEDED(AudioEngine.Initialize(&config))) {
             // create audio clip
             auto clip = WrapAL::CreateAudioClip(
-                WrapAL::AudioFormat::Format_OggVorbis,
+                WrapAL::EncodingFormat::Format_OggVorbis,
                 //L"NationalAnthemOfRussia.ogg",
                 L"Sakura.ogg",
                 WrapAL::Flag_StreamingReading